% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fs_pkg.R
\name{create_fs_pkg}
\alias{create_fs_pkg}
\title{Create an R package based on the Farming Stats template}
\usage{
create_fs_pkg()
}
\arguments{
\item{pkg_name}{string containg package name.}
}
\value{
The output will be all the necessary package
files within your repo.
}
\description{
This function will create an R package using the Farming
Stats template.
}
\details{
Use this function to create an R package based on the
farming stats template. The function should be run
within an already established R project.

For easy git integration, first set up a repo on
GitHub, then clone the repo to you local directory,
then run this function. Though this is not necessary
if your package is not being pushed to GitHub.

This function will add:
\itemize{
\item DESCRIPTION
\item NAMESPACE
\item MIT License
\item Defra licence
\item R folder
\item gitignore
\item buildignore
\item README Qmd and MD
}
}
