% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fs_script.R
\name{create_fs_script}
\alias{create_fs_script}
\title{Create a new script with farming stats header.}
\usage{
create_fs_script(
  file_name = NULL,
  folder = NULL,
  author = NULL,
  email = NULL,
  date = format(Sys.Date(), "\%d/\%m/\%Y")
)
}
\arguments{
\item{file_name}{string containing desired name
for script.}

\item{folder}{string containing folder name to
save script. This is built on the \code{here()} function
in R, so will follow your root directory. If you
want to save in a sub-folder, enter the full folder
sequence, e.g. "folder/sub-folder".}

\item{author}{string containing author's name.}

\item{email}{string containing author's email.}

\item{date}{string containing a date. By default,
this will be set as today's date.}
}
\value{
An R script will be saved in the root directory
or in the specified folder.
}
\description{
This function will create a new script with
the farming stats header added. There a multiple
options for cutomisation (see details).
}
\details{
This is a simple function to create a new script
with the farming stats header template added.

By just calling the function a script will be
created with the template entered. All fields of
the template will be blank except for the date,
which is set to today's date.

There are options for customising and entering
some details automatically. The details that can
be pre-entered are:
#' \itemize{
\item script name
\item location to save
\item author
\item author's email
\item date created
}

By default these are set to NULL (except for the
date). But can be set to any string.
}
