% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_connect_ssh.R
\name{fs_connect_ssh}
\alias{fs_connect_ssh}
\title{Connect RStudio to your GitHub account using a SSH}
\usage{
fs_connect_ssh(username, email)
}
\arguments{
\item{username}{user's GitHub username}

\item{email}{user's email address}
}
\description{
Connect RStudio to your GitHub account using SSH - this is the
preferred method of connecting RStudio and GitHub on the DASH platform. The
use will only need to supply their GitHub username and email address, then
the function will set their credentials and generate an SSH key. Once the
SSH key has been added to GitHub, the function does the final bits of
set-up. All steps and outputs are printed in the RStudio console rather
than the terminal, which offers a more user friendly experience.
}
\details{
The suggested method for connecting RStudio and GitHub on the DASH
platform is via SSH. The alternative is via Personal Access Token, but this
method can be a bit clunky and frustrating on the DASH platform. With SSH
providing a much smoother user experience.

This function is designed to simplify the connection process and avoid the
user having to interact with the terminal - which many beginners to Git
find intimidating. Instead, all prompts and outputs are returned in the
RStudio console.

The users calls the function and gives their user credentials (username and
email), the function will set these and print an output to confirm they
have been set. The function will then create a hidden SSH folder and
sub-folder for "id_ed25519" (~/.ssh/id_ed25519). Once done, it will then
generate your SSH key and save it to this folder. The SSH key will then be
printed in the console. The function then adds GitHub as a known host.

At this point, the function pauses and asks the user to confirm they have
added the SSH key to GitHub. The user can then copy the SSH key from the
console, got to GitHub -> settings -> SSH and GPG keys and add the SSH key.

Once added, the user can respond to the prompt and the function will finish
establishing the connection. If all works as expected, the message "Hi
username! You've successfully authenticated, but GitHub does not provide
shell access" will be printed in the console.
}
\author{
Josh Moatt
}
