% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_proj.R
\name{fs_proj}
\alias{fs_proj}
\title{Title}
\usage{
fs_proj(path, ...)
}
\arguments{
\item{argument}{description add more as needed}
}
\value{
what is the output
}
\description{
Description of function
}
\details{
Details of function (optional)
}
