% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_use_github.R
\name{fs_use_github}
\alias{fs_use_github}
\title{Create a GitHub repository from local project.}
\usage{
fs_use_github(message = "Initial commit", private = TRUE)
}
\arguments{
\item{message}{initial commit message. Default
is "Initial commit".}

\item{private}{if TRUE creates private repo.}
}
\value{
R project is truned into a git repo and an
associated github repo si created in the users
github account.
}
\description{
This simple function will turn your local R
project into a git repo, then create a repo
on GitHub and perform the initial set up and
commit.

Note: by default creates a private repo
to ensure security. Can be change if the
repo needs to be public. Can also be done
in GitHub at a later date.
}
