% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_databricks_pat.R
\name{set_databricks_pat}
\alias{set_databricks_pat}
\title{Set databricks PAT}
\usage{
set_databricks_pat()
}
\value{
databricks PAT written to environment
}
\description{
Simple function to set databricks PAT. Done via function to avoid committing to GitHub
accidentally.
}
\author{
Josh Moatt
}
