% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_file.R
\name{temp_file}
\alias{temp_file}
\title{function for creating a tempfile when exporting data to DASH}
\usage{
temp_file(object, file_type)
}
\arguments{
\item{object}{object from R environment to export. So far works with all
types (e.g. data.frame, openxlsx workbook etc).}

\item{file_type}{string  specifying file type of the tempfile. Can be "rds",
"xlsx" or "csv". Other types can be added to the function if needed.}
}
\value{
a tempfile with specified extension.
}
\description{
Very simple function to aid in exporting an object from the R environment on
a DASH RStudio cluster to the databricks filestore. Desingned to be used
alongside the function {db_volume_write} from the \code{brickster} package.
}
\author{
Josh Moatt
}
