% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_volume_get.R
\name{uc_volume_get}
\alias{uc_volume_get}
\title{Pull data from the DASH Unity Catalog into an RStudio cluster}
\usage{
uc_volume_get(workspace, volume, token, out_file)
}
\arguments{
\item{workspace}{databricks workspace (string)}

\item{volume}{file path to data folder (string)}

\item{token}{databricks PAT (string)}

\item{out_file}{filepath on RStudio Cluster that data should be stored.}
}
\value{
data saved in Rstudio cluster.
}
\description{
Currently the DASH Unity Catalog (UC) and RStudio server cannot directly
interact. So you must interact with the databricks API and copy data from
DASH UC to the RStudio server. This is a simple function to read data from UC
into RStudio cluster on databricks.

Note, you must have set up a databricks PAT in the databricks settings before
doing this.
}
\author{
Josh Moatt
}
