% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_volume_put.R
\name{uc_volume_put}
\alias{uc_volume_put}
\title{Write data to the DASH Unity Catalog from RStudio cluster}
\usage{
uc_volume_put(workspace, volume, file, token, folder)
}
\arguments{
\item{workspace}{databricks workspace (string)}

\item{volume}{path to folder on databricks where data should be written
(string)}

\item{token}{databricks PAT (string)}

\item{folder}{path to folder on RStudio Cluster that contains data to export.}
}
\value{
data saved in Rstudio cluster.
}
\description{
Currently the DASH Unity Catalog (UC) and RStudio server cannot directly
interact. So you must interact with the databricks API and copy data to the
DASH UC from the RStudio server. This is a simple function to save data from
the RStudio cluster onto databricks. It is built using the httr2 package and
uses a PUT request to save the data.

Note, you must have set up a databricks PAT in the databricks settings before
doing this.
}
\author{
Josh Moatt
}
