% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_proj.R
\name{fs_proj}
\alias{fs_proj}
\title{Function for RStudio project template}
\usage{
fs_proj(path, ...)
}
\description{
This is a function that is called in the "New project"
viewer pane when the user chooses a Farming Stats
Project template. It should not be used away from the
RStudio viewer. If you would like to manually add project
files, use the function \code{\link{create_fs_proj}} instead.

I have not included additional information on how to use
this function, as it is not intended to be used outside
the template call.

To subsequently link this to a github repo, the best
plan is to use \code{\link{fs_use_github}}.
}
